move the filter trait implementations to more obvious place
